extract a mixin for the listeners
