+++ lib/shouter/listeners/async.rb
@@ -1,15 +1,10 @@
 module Shouter
   module Listeners
     class Async
+
+      include Shouter::Mixin
+
       attr_reader :object, :options, :scope
-
-      def initialize(object, options)
-        raise Shouter::ScopeMissingError unless options[:scope]
-
-        @object = object
-        @options = options
-        @scope = options[:scope]
-      end
 
       def notify(scope, event, args, &block)
         return unless notification_allowed?(event, scope)
@@ -21,33 +16,6 @@ module Shouter
           Store.unregister(object) if single?
         end
       end
-
-      private
-
-      def notification_allowed?(event, desired_scope)
-        object.respond_to?(event) && scope == desired_scope
-      end
-
-      def fire_hook!(callback)
-        Shouter::Hook.(callback)
-      end
-
-      def fire_guard!
-        Shouter::Guard.(guard)
-      end
-
-      def callback
-        options[:callback]
-      end
-
-      def single?
-        options[:single] == true
-      end
-
-      def guard
-        options[:guard]
-      end
     end
   end
 end
-
